refactor: Remove unused method-related files and interfaces
The code changes remove the unused method-related files and interfaces, including method.service.interface.ts and method.base.interface.ts. These files were no longer being used in the codebase and were causing unnecessary clutter. Removing them improves code maintainability and reduces potential confusion.
