+++ src/modules/method/application/interface/method.base.interface.ts
@@ -1,0 +1,12 @@
+export interface IMethodValues {
+  name: string;
+  inputs: { name: string; type: string }[];
+  outputs: { type: string }[];
+  params?: { name: string; value: string }[];
+  docs: string;
+  invocationId: string;
+}
+
+export interface IUpdateMethodValues extends Partial<IMethodValues> {
+  id: string;
+}
